Add files via upload
These are the resultant csv files that were generated after running the Halliday python script. The first is a list of all hyperlinks scraped from the website, and the second is the result of the hyperlinks being passed into the script as a list. The individual pages were then scraped individually and appended to the main csv.